fix: allow float duration in VideoMetadata model
- Change duration field from int to float to handle precise video durations
- This fixes Pydantic validation errors when yt-dlp returns fractional seconds
- Twitter/X videos often have precise durations like 27.79 seconds

This resolves the validation error and allows successful metadata extraction.

+++ video-processor/main.py
@@ -45,7 +45,7 @@ class VideoMetadataRequest(BaseModel):
 
 class VideoMetadata(BaseModel):
     title: str
-    duration: int  # in seconds
+    duration: float  # in seconds (allow float for precise duration)
     thumbnail: str
     uploader: str
     upload_date: str
